[CELEBORN-21][REFACTOR] Extract revive related logical from LifecycleManager
